requirements.txt
# Core dependencies
pandas==2.3.2
numpy==1.26.4
matplotlib==3.10.8
ipykernel==7.1.0
pyyaml==6.0.3

# Machine Learning
scikit-learn==1.8.0
xgboost==3.1.3